requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
flask==3.1.2
flask-cors==5.0.0
Pillow==10.2.0
pillow-avif-plugin==1.4.3
